something interesing
A loop to detect the inpunt number,if the input  is equal to the
"ff",the led will be on.When the next time it is not detexted at
"ff",the led will be off.
There is a count in 1ms loop,every loop the count add 1.When the input
is "ff",the count will be 0,and the flag=1.When the inpit is not
"ff",led off and print the count.So we can see the system time.
But this way do not have same result you give us.
I am confused.